:ok_hand: Crud Realizado!!

+++ gerenciamentoDeClientes/api/src/main/java/com/develop/gpp/domain/repository/ClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<ClienteModel, Integer>{
+public interface ClienteRepository extends JpaRepository<ClienteModel, Long>{
 
 }
